refactor: moved contract validators to contracts folder
